requirements.txt
googletrans==4.0.0-rc1
langdetect==1.0.9
python-frontmatter==1.0.1